fix: improve files architecture
