+++ src/models/parameters/processing/bounding_box_cropper_parameters.py
