+++ src/models/steps/data_validation/common/object_detection_dataset_context_validator.py
@@ -1,4 +1,4 @@
-from src.steps.data_validation.utils.dataset_context_validator import (
+from src.models.steps.data_validation.common.dataset_context_validator import (
     DatasetContextValidator,
 )
