+++ src/models/steps/processing/bounding_box_cropper_processing.py
@@ -1,19 +1,12 @@
 import os
 
-import cv2
-import numpy as np
+from PIL import Image
 from picsellia import DatasetVersion, Client
 from picsellia.types.enums import TagTarget, InferenceType
 from picsellia_annotations.coco import Annotation
 
-from src import Pipeline
-from src import step
-from src.models.contexts.picsellia_context import PicselliaProcessingContext
 from src.models.dataset.dataset_context import DatasetContext
-from src.pipelines.bounding_box_cropper.utils.bounding_box_cropper_parameters import (
-    BoundingBoxCropperParameters,
-)
-from src.steps.processing.utils.dataset_version_creation_processing import (
+from src.models.steps.processing.dataset_version_creation_processing import (
     DatasetVersionCreationProcessing,
 )
 
@@ -70,7 +63,7 @@ class BoundingBoxCropperProcessing(DatasetVersionCreationProcessing):
             image_filename (str): The filename of the image to process.
         """
         image_filepath = os.path.join(self.dataset_context.image_dir, image_filename)
-        image = cv2.imread(image_filepath)
+        image = Image.open(image_filepath)
         coco_files_image_ids = [
             coco_file_image.id
             for coco_file_image in self.dataset_context.coco_file.images
@@ -97,22 +90,22 @@ class BoundingBoxCropperProcessing(DatasetVersionCreationProcessing):
                     )
 
     def _extract(
-        self, image: np.ndarray, image_filename: str, coco_file_annotation: Annotation
+        self, image: Image, image_filename: str, coco_file_annotation: Annotation
     ) -> None:
         """
         Extracts the bounding box from the image and saves it to the processed dataset directory.
 
         Args:
-            image (np.ndarray): The image to extract the bounding box from.
+            image (Image): The image to extract the bounding box from.
             image_filename (str): The filename of the image.
             coco_file_annotation (Annotation): The annotation containing the bounding box.
         """
-        extracted_image = image[
-            int(coco_file_annotation.bbox[1]) : int(coco_file_annotation.bbox[1])
-            + int(coco_file_annotation.bbox[3]),
-            int(coco_file_annotation.bbox[0]) : int(coco_file_annotation.bbox[0])
-            + int(coco_file_annotation.bbox[2]),
-        ]
+        x = int(coco_file_annotation.bbox[0])
+        y = int(coco_file_annotation.bbox[1])
+        width = int(coco_file_annotation.bbox[2])
+        height = int(coco_file_annotation.bbox[3])
+
+        extracted_image = image.crop((x, y, x + width, y + height))
 
         label_folder = os.path.join(
             self.processed_dataset_context.image_dir, self.label_name_to_extract
@@ -122,7 +115,7 @@ class BoundingBoxCropperProcessing(DatasetVersionCreationProcessing):
         processed_image_filename = f"{os.path.splitext(image_filename)[0]}_{self.label_name_to_extract}_{coco_file_annotation.id}.{image_filename.split('.')[-1]}"
         processed_image_filepath = os.path.join(label_folder, processed_image_filename)
 
-        cv2.imwrite(processed_image_filepath, extracted_image)
+        extracted_image.save(processed_image_filepath)
 
     def _add_processed_images_to_dataset_version(self) -> None:
         """
@@ -153,20 +146,3 @@ class BoundingBoxCropperProcessing(DatasetVersionCreationProcessing):
         """
         self._process_images()
         self._add_processed_images_to_dataset_version()
-
-
-@step
-def bounding_box_cropper_processing(dataset_context: DatasetContext):
-    context: PicselliaProcessingContext[
-        BoundingBoxCropperParameters
-    ] = Pipeline.get_active_context()
-
-    processor = BoundingBoxCropperProcessing(
-        client=context.client,
-        input_dataset_context=dataset_context,
-        label_name_to_extract=context.processing_parameters.label_name_to_extract,
-        output_dataset_version=context.output_dataset_version,
-        datalake=context.processing_parameters.datalake,
-        destination_path=os.path.join(os.getcwd(), str(context.job_id)),
-    )
-    processor.process()
